Partial mocks now use _objc_msgForward_stret for the forwarding IMP, as appropriate.
Without using the stret variant to match a method type encoding
which specifies the return of a struct, class_addMethod may silently corrupt
the method lookup table for the subclass, causing the partial-mocking
of the specified method to fail and an exception to be thrown when
subclass is next messaged (with _any_ message).

A test has been added, which test fails without this commit's changes.