OCMockClassObject now specifically adds support for partially mocking class objects.
It has been renamed to OCPartialMockClassObject, and now implements
an OCPartialMock interface shared with OCPartialMockObject.

This allows it to leverage the existing functionality defined by the
OCPartialMockRecorder and OCMRealObjectForwarder, and clarifies
its function.

Support for purely mocking class objects, i.e. an analogue of OCClassMockObject
for mocking class methods rather than instance methods,
will be added in a subsequent commit.