Before setting up a forwarding imp for a partial mock class object, retrieve the original IMP via another method than class_replaceMethod.
class_replaceMethod is described to return "The previous implementation of the method
identified by name for the class identified by cls." However, it turns out that
it only considers the previous implementation *within the class* -- not its superclasses.
class_replaceMethod will thus return NULL if the class being mocked doesn't itself
define the mocked method, but rather its superclass.

This will lead to us adding back the original method (under our prefixed name)
with a NULL IMP.

method_getImplementation, on the other hand, returns the IMP of the method
on the superclass.