Added support for partial class mocks.
Partial class mocks allow functionality to be stubbed out across
ALL instances of a class. They are to be used, instead of partial object mocks,
where the identities of individual instances are unknown or unimportant.

They differ from the other partial object mock types, and from class mocks,
in that they cannot be used directly, as instances of the mocked class.
This is because there is no single "real object" to which to forward unrecorded
invocations and "andForwardToRealObject"-handled invocations from a
directly-used mock. For this reason, OCPartialClassMockObject does not
descend from OCClassMockObject.