Add an "any selector" option to OCMArg.
Although SELs are typedef'd pointers, the -anyPointer option could not
be used because NSInvocation+OCMAdditions would try to get
the selector object-value as a string--which would be nil, for a pointer which
wasn't a selector. I end up changing NSInvocation+OCMAdditions
to box SELs as pointers anyway, but I figured a more explicit interface
to OCMArg than -anyPointer couldn't hurt.