Add +[OCMArg anyCString] for parameters with char * type
Adds support for stubbing of methods that take char * parameters.

+[OCMArg anyPointer] doesn't work in the case of char * pointers because
NSInvocation retains the C string argument by copying its contents and
referencing the new copy. This new copy will have a different memory address
than the original string and will break matching of +[OCMArg anyPointer].
This is because +[OCMArg anyPointer] is defined as (void *)0x01234567 and
the matching process expects the argument of the recorded method to remain
unchanged.

The solution is to allow NSInvocation to copy the string and to later match
the stub method by comparing the contents of the string. This is potentially
unsafe because reads 8 bytes from whatever address the argument points to.
However, this should generally work in practice if all the pointers are ones
returned by malloc() because Apple's implemetation of malloc() returns 16-byte
aligned objects.