Resolve BOOL vs. char type bug
Updates the type checking in OCMBoxedReturnValueProvider's
handleInvocation method so it's closer to the system used
by the main OCMock project on TOT.  The new implementation
uses a helper method, -[NSValue getBytes:objCType:], which
treates 'B' and 'c' types as equivalent.

Note: NSValue+OCMAdditions and the meat of the new implementation
of handleInvocation are both copied from the main OCMock project
(at revision erikdoe/ocmock@999b843).  Anyone looking for a useful
history of this code will need to search in the main OCMock project.