Fix OCMock boxed value unit tests
We no longer necessarily raise an exception when boxed value types
don't match.  If the types can be converted losslessly then we do
not raise an exception.  This commit changes the old testRaises*
test to attempt a stub and return with actually incompatible types
and adds a new test, testDoesNotRaise*, that does a stub and return
with different but compatible types, verifying that no exception is
raised.