64-bit cleanup: use %tu & %zd for formats
while the officially supported Apple position is to cast the value,
there are some circumstances in which this can result in a different
value being printed than is held.  this can be particularly troublesome
for values that are meant to be held as 64-bit IDs.

this solution is pretty widely accepted as a way to have a single
specifier for each solution that works on both 32-bit & 64-bit archs.